aus-cli dump commands don't require an OCM session for
this PR removes the need to be logged into OCM when using a commands `--dump` option. for commands that offer patch or replace functionality, an active OCM login is still required when used in patch mode. for a true gitops way of handling AUS configuration, `--replace` is the way to go anyways, so this should not be too restrictive.

part of https://issues.redhat.com/browse/APPSRE-8707

Signed-off-by: Gerd Oberlechner <goberlec@redhat.com>